move chapter table sort/filters into recoil selector
